Add Navigation interaction
