+++ app/AppNavigator.js
@@ -1,17 +1,21 @@
 'use strict'
 import React, { Component } from 'react';
-import { StyleSheet, View, Image, Navigator } from 'react-native';
-import DetailView from '../app/Views/DetailView'
+import { StyleSheet, View, Navigator, Text } from 'react-native';
+import { DetailView } from '../app/Views/DetailView'
+import { TeamsListView } from '../app/Views/TeamsListView'
 
 class AppNavigator extends Component {
 
   _renderScene(route, navigator) {
     var globalNavigatorProps = { navigator }
-
     switch(route.ident) {
+      case "TeamsListView":
+      return (
+        <TeamsListView  navigator={navigator} test="HOLA MUNDO"/>
+      )
       case "DetailView":
         return (
-          <DetailView {...globalNavigatorProps} />
+          <DetailView  navigator={navigator}/>
         )
       default:
         return (
@@ -25,13 +29,11 @@ class AppNavigator extends Component {
       <Navigator
         initialRoute={this.props.initialRoute}
         ref="appNavigator"
-        style={styles.navigatorStyles}
         renderScene={this._renderScene}
         configureScene={(route) => ({
-          ...route.sceneConfig || Navigator.SceneConfigs.FloatFromRight })} />
+          ...route.sceneConfig || Navigator.SceneConfigs.FloatFromRight })}/>
     )
   }
-
 }
 
 const styles = StyleSheet.create({
